+++ packages/0xsequence/src/utils.ts
@@ -1,5 +1,12 @@
 export * from '@0xsequence/utils'
 
+export {
+  isValidSignature,
+  isValidMessageSignature,
+  isValidTypedDataSignature,
+  recoverWalletConfig
+} from '@0xsequence/provider'
+
 export type {
   Deferrable,