+++ packages/indexer/src/index.ts
@@ -4,9 +4,17 @@ import fetch from 'cross-fetch'
 
 import { Indexer as BaseSequenceIndexer } from './indexer.gen'
 
+export enum SequenceIndexerServices {
+  MAINNET = 'https://mainnet-indexer.sequence.app',
+  POLYGON = 'https://polygon-indexer.sequence.app',
+  
+  RINKEBY = 'https://rinkeby-indexer.sequence.app',
+  POLYGON_MUMBAI = 'https://mumbai-indexer.sequence.app'
+}
+
 export class SequenceIndexerClient extends BaseSequenceIndexer {
   constructor(hostname: string, public jwtAuth?: string) {
-    super(hostname, fetch)
+    super(hostname.endsWith('/') ? hostname.slice(0, -1) : hostname, fetch)
     this.fetch = this._fetch
   }