+++ packages/provider/src/transports/wallet-request-handler.ts
@@ -311,7 +311,7 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
           } else {
             const promptResultForDeployment = await this.handleConfirmWalletDeployPrompt(this.prompter, signer, chainId)
             if (promptResultForDeployment) {
-              sig = await this.prompter.promptSignMessage({ chainId: chainId, message: message })
+              sig = await this.prompter.promptSignMessage({ chainId: chainId, message: message }, this.connectOptions)
             }
           }
 
@@ -351,7 +351,7 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
           } else {
             const promptResultForDeployment = await this.handleConfirmWalletDeployPrompt(this.prompter, signer, chainId)
             if (promptResultForDeployment) {
-              sig = await this.prompter.promptSignMessage({ chainId: chainId, typedData: typedData })
+              sig = await this.prompter.promptSignMessage({ chainId: chainId, typedData: typedData }, this.connectOptions)
             }
           }
 
@@ -375,7 +375,7 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
             txnHash = txnResponse.hash
           } else {
             // prompt user to provide the response
-            txnHash = await this.prompter.promptSendTransaction(transactionParams, chainId)
+            txnHash = await this.prompter.promptSendTransaction(transactionParams, chainId, this.connectOptions)
           }
 
           if (txnHash) {
@@ -404,7 +404,7 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
             // we will want to resolveProperties the bignumber values to hex strings
             response.result = await signer.signTransactions(transaction, chainId)
           } else {
-            response.result = await this.prompter.promptSignTransaction(transaction, chainId)
+            response.result = await this.prompter.promptSignTransaction(transaction, chainId, this.connectOptions)
           }
 
           break
@@ -482,6 +482,23 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
         case 'eth_gasPrice': {
           const gasPrice = await provider.getGasPrice()
           response.result = gasPrice.toHexString()
+          break
+        }
+
+        case 'wallet_switchEthereumChain': {
+          const [switchParams] = request.params!
+          if (!switchParams.chainId || switchParams.chainId.length === 0) {
+            throw new Error('invalid chainId')
+          }
+
+          const chainId = ethers.BigNumber.from(switchParams.chainId)
+
+          const ok = await this.setDefaultNetwork(chainId.toString(), true)
+          if (!ok) {
+            throw new Error(`unable to set chainId ${chainId}`)
+          }
+
+          response.result = null // success
           break
         }
 
@@ -774,7 +791,7 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
       return true
     }
 
-    const promptResult = await prompter.promptConfirmWalletDeploy(chainId)
+    const promptResult = await prompter.promptConfirmWalletDeploy(chainId, this.connectOptions)
     // if client returned true, check again to make sure wallet is deployed and up to date
     if (promptResult) {
       if (!await signer.isSettled(chainId)) {
@@ -789,11 +806,11 @@ export class WalletRequestHandler implements ExternalProvider, JsonRpcHandler, P
 }
 
 export interface WalletUserPrompter {
-  promptConfirmWalletDeploy(chainId: number): Promise<boolean>
   promptConnect(options?: ConnectOptions): Promise<PromptConnectDetails>
-  promptSignMessage(message: MessageToSign): Promise<string>
-  promptSignTransaction(txn: TransactionRequest, chaindId?: number): Promise<string>
-  promptSendTransaction(txn: TransactionRequest, chaindId?: number): Promise<string>
+  promptSignMessage(message: MessageToSign, options?: ConnectOptions): Promise<string>
+  promptSignTransaction(txn: TransactionRequest, chaindId?: number, options?: ConnectOptions): Promise<string>
+  promptSendTransaction(txn: TransactionRequest, chaindId?: number, options?: ConnectOptions): Promise<string>
+  promptConfirmWalletDeploy(chainId: number, options?: ConnectOptions): Promise<boolean>
 }
 
 const permittedJsonRpcMethods = [