+++ packages/relayer/src/index.ts
@@ -1,5 +1,5 @@
 import { ethers, providers } from 'ethers'
-import { TransactionBundle, Transaction } from '@0xsequence/transactions'
+import { TransactionBundle, Transaction, TransactionResponse } from '@0xsequence/transactions'
 import { WalletContext } from '@0xsequence/network'
 import { WalletConfig } from '@0xsequence/config'
 import { proto } from './rpc-relayer'
@@ -16,11 +16,15 @@ export interface Relayer {
     ...transactions: Transaction[]
   ): Promise<Transaction[]>
 
+  // getFeeOptions returns the fee options that the relayer will accept as payment.
+  // If a quote is returned, it may be passed back to the relayer for dispatch.
+  getFeeOptions(
+    bundle: TransactionBundle
+  ): Promise<{ options: FeeOption[], quote?: FeeQuote }>
+
   // gasRefundOptions returns the transactions which can be included to refund a
   // relayer for submitting your transaction to a network.
   gasRefundOptions(
-    config: WalletConfig,
-    context: WalletContext,
     bundle: TransactionBundle
   ): Promise<FeeOption[]>
 
@@ -30,10 +34,10 @@ export interface Relayer {
   getNonce(config: WalletConfig, context: WalletContext, space?: ethers.BigNumberish, blockTag?: providers.BlockTag): Promise<ethers.BigNumberish>
 
   // relayer will submit the transaction(s) to the network and return the transaction response.
-  relay(signedTxs: TransactionBundle): Promise<providers.TransactionResponse>
+  relay(signedTxs: TransactionBundle, quote?: FeeQuote): Promise<TransactionResponse>
 
   // wait for transaction confirmation
-  wait(metaTxnId: string | TransactionBundle, timeout: number): Promise<providers.TransactionResponse>
+  wait(metaTxnId: string | TransactionBundle, timeout: number): Promise<TransactionResponse>
 }
 
 export * from './local-relayer'
@@ -43,7 +47,26 @@ export { proto as RpcRelayerProto } from './rpc-relayer'
 export type SimulateResult = proto.SimulateResult
 export type FeeOption = proto.FeeOption
 
+// A fee quote is simply an opaque value that can be obtained via Relayer.getFeeOptions(), and
+// returned back to the same relayer via Relayer.relay(). Fee quotes should be treated as an
+// implementation detail of the relayer that produces them.
+//
+// This interface exists for type-safety purposes to protect against passing non-FeeQuotes to
+// Relayer.relay(), or any other functions that call it indirectly (e.g. Account.sendTransaction).
+export interface FeeQuote {
+  _tag: 'FeeQuote'
+  _quote: unknown
+}
+
 export function isRelayer(cand: any): cand is Relayer {
-  return cand && cand.estimateGasLimits !== undefined && cand.gasRefundOptions !== undefined &&
-    cand.getNonce !== undefined && cand.relay !== undefined
+  return (
+    typeof cand === 'object' &&
+    typeof cand.simulate === 'function' &&
+    typeof cand.estimateGasLimits === 'function' &&
+    typeof cand.getFeeOptions === 'function' &&
+    typeof cand.gasRefundOptions === 'function' &&
+    typeof cand.getNonce === 'function' &&
+    typeof cand.relay === 'function' &&
+    typeof cand.wait === 'function'
+  )
 }