+++ packages/utils/src/sleep.ts
@@ -1,5 +1,3 @@
-import { clearConfigCache } from "prettier"
-
 export const sleep = (t: number) => {
   return new Promise<void>(resolve => {
     const timeout = setTimeout(() => {