+++ packages/wallet/tests/wallet.spec.ts
@@ -38,7 +38,7 @@ const HookCallerMockArtifact = require('@0xsequence/wallet-contracts/artifacts/c
 const Web3 = require('web3')
 const { expect } = chai.use(chaiAsPromised)
 
-configureLogger({ logLevel: 'DEBUG' })
+configureLogger({ logLevel: 'DEBUG', silence: false })
 
 import hardhat from 'hardhat'
 import { Interface } from 'ethers/lib/utils'